+++ openfisca_france/model/revenus/capital/plus_value.py
@@ -2,16 +2,32 @@
 
 from openfisca_france.model.base import *  # noqa analysis:ignore
 
-# Gain de levée d'options
-# Bouvard: j'ai changé là mais pas dans le code, il faut chercher les f1uv
-# et les mettre en f1tvm comme pour salaire_imposable
-# Il faut aussi le faire en amont dans les tables
+## PLAN :
+# 1) Gains taxés comme des salaires
+# 2) Autres plus-values (exonérées, au barème, ...)
+# 3) Abattements sur plus-values
+# 4) Plus-values taxées forfaitairement
+# 5) Autres variables
 
-# là je ne comprends pas pourquoi il faut changer les f1uv en f1tvm....
-# du coups je n'ai pas changé et j'ai fait un dico comme pour salaire_imposable
+
+# Gains taxés comme des salaires
+
+class f1tt(Variable):
+    cerfa_field = {
+        0: u"1TT",
+        1: u"1UT",
+        }
+    value_type = int
+    unit = 'currency'
+    entity = Individu
+    label = u"Gains de levée d'options sur titres et gains d'acquisition d'actions gratuites attribuées à compter du 28.9.2012"
+    # start_date = date(2012, 1, 1)
+    definition_period = YEAR
+
 
 class f1tv(Variable):
-    cerfa_field = {0: u"1TV",
+    cerfa_field = {
+        0: u"1TV",
         1: u"1UV",
         }
     value_type = int
@@ -21,10 +37,10 @@ class f1tv(Variable):
     end = '2015-12-31'
     definition_period = YEAR
 
-  # (f1tv,f1uv))
 
 class f1tw(Variable):
-    cerfa_field = {0: u"1TW",
+    cerfa_field = {
+        0: u"1TW",
         1: u"1UW",
         }
     value_type = int
@@ -34,10 +50,10 @@ class f1tw(Variable):
     end = '2015-12-31'
     definition_period = YEAR
 
-  # (f1tw,f1uw))
 
 class f1tx(Variable):
-    cerfa_field = {0: u"1TX",
+    cerfa_field = {
+        0: u"1TX",
         1: u"1UX",
         }
     value_type = int
@@ -47,44 +63,69 @@ class f1tx(Variable):
     end = '2016-12-31'
     definition_period = YEAR
 
-  # (f1tx,f1ux))
 
-
-
-class f3si(Variable):
+class f1tz(Variable):
+    cerfa_field = u"1TZ"
     value_type = int
-    entity = FoyerFiscal
-    # start_date = date(2012, 1, 1)
-    definition_period = YEAR
-
-  # TODO: parmi ces cas créer des valeurs individuelles
-#                                    # correspond à autre chose en 2009, vérifier 2011,2010
-
-class f3sa(Variable):
-    cerfa_field = u"3SA"
-    value_type = int
-    entity = FoyerFiscal
-    label = u"Plus-values de cessions de titres réalisées par un entrepreneur, taxables à 19%"
-    # start_date = date(2012, 1, 1)
-    end = '2012-12-31'
+    unit = 'currency'
+    entity = Individu
+    label = u"Gains d'acquisition d'actions gratuites attribuées sur décision prise à compter du 8.8.2015"
+    # start_date = date(2015, 1, 1)
     definition_period = YEAR
 
 
-class f3sf(Variable):
+class f3vj(Variable):
+    cerfa_field = {
+        0: u"3VJ",
+        1: u"3VK",
+        }
     value_type = int
-    entity = FoyerFiscal
-    # start_date = date(2012, 1, 1)
+    unit = 'currency'
+    entity = Individu
+    label = u"Gains imposables sur option dans la catégorie des salaires"
     definition_period = YEAR
 
-  # TODO: déjà définit plus haut, vérifier si 2009, 2010, 2011 correspondent à la même chose que 12 et 13
 
-class f3sd(Variable):
+# Autres plus-values
+
+class f3vg(Variable):
+    cerfa_field = u"3VG"
     value_type = int
+    unit = 'currency'
     entity = FoyerFiscal
-    # start_date = date(2012, 1, 1)
+    label = u"Plus-value imposable sur gains de cession de valeurs mobilières, de droits sociaux et gains assimilés"
     definition_period = YEAR
 
-  # TODO: déjà définit plus haut, vérifier si 2009, 2010, 2011 correspondent à la même chose que 12 et 13
+
+class f3vh(Variable):
+    cerfa_field = u"3VH"
+    value_type = int
+    unit = 'currency'
+    entity = FoyerFiscal
+    label = u"Moins-value imposable su gains de cession de valeurs mobilières, de droits sociaux et gains assimilés"
+    definition_period = YEAR
+
+
+class f3vv(Variable):
+    cerfa_field = u"3VV"
+    value_type = int
+    unit = 'currency'
+    entity = FoyerFiscal
+    label = u"Plus-values réalisées par les non-résidents: montant du prélèvement de 45 % déjà versé"
+    # start_date = date(2013, 1, 1)
+    end = '2013-12-31'
+    definition_period = YEAR
+
+
+class f3vz(Variable):
+    cerfa_field = u"3VZ"
+    value_type = int
+    unit = 'currency'
+    entity = FoyerFiscal
+    label = u"Plus-values imposables sur cessions d’immeubles ou de biens meubles"
+    # start_date = date(2011, 1, 1)
+    definition_period = YEAR
+
 
 class f3vc(Variable):
     cerfa_field = u"3VC"
@@ -103,6 +144,25 @@ class f3vp(Variable):
     entity = FoyerFiscal
     label = u"Plus-values exonérées de cessions de titres de jeunes entreprises innovantes"
     # start_date = date(2007, 1, 1)
+    end = '2013-12-31'
+    definition_period = YEAR
+
+
+class f3vq(Variable):
+    cerfa_field = u"3VQ"
+    value_type = int
+    unit = 'currency'
+    entity = FoyerFiscal
+    label = u"Cessions de titres détenus à l'étranger par les impatriés : plus-values exonérées (50%)"
+    definition_period = YEAR
+
+
+class f3vr(Variable):
+    cerfa_field = u"3VR"
+    value_type = int
+    unit = 'currency'
+    entity = FoyerFiscal
+    label = u"Cessions de titres détenus à l'étranger par les impatriés : moins-values non imputables (50%)"
     definition_period = YEAR
 
 
@@ -113,21 +173,9 @@ class f3vy(Variable):
     entity = FoyerFiscal
     label = u"Plus-values exonérées de cessions de participations supérieures à 25% au sein du groupe familial"
     # start_date = date(2011, 1, 1)
+    end = '2013-12-31'
     definition_period = YEAR
 
-
-class f3vd(Variable):
-    cerfa_field = {0: u"3VD",
-        1: u"3SD",
-        }
-    value_type = int
-    unit = 'currency'
-    entity = Individu
-    label = u"Gains de levée d'options sur titres et gains d'acquisition d'actions taxables à 18 %"
-    # start_date = date(2008, 1, 1)
-    definition_period = YEAR
-
-  # (f3vd, f3sd)
 
 class f3ve(Variable):
     cerfa_field = u"3VE"
@@ -138,95 +186,133 @@ class f3ve(Variable):
     definition_period = YEAR
 
 
-class f3vf(Variable):
-    cerfa_field = {0: u"3VF",
-        1: u"3SF",
-        }
-    value_type = int
-    unit = 'currency'
-    entity = Individu
-    label = u"Gains de levée d'options sur titres et gains d'acquisition d'actions taxables à 41 %"
-    definition_period = YEAR
-
-  # (f3vf, f3sf)
-
-# comment gérer les cases qui ont le même nom mais qui ne correspondent pas tout à fait à la même chose ?
-# peut-ont garder le même nom et l'encadrer par des start-end ? ou avec un truc genre if sur l'année ?(pour ne pas avoir à changer le nom de la variable)
-# si on garde le même nom avec des start-end, et si on intégre la variable partout où elle doit être (dans les différents calculs), est-on sûr que lors des calculs les start-end seront bien pris en compte ?
-# ça rendra le modéle un peu moins clair parce qu'il y aura le même nom de variable pour des choses différentes et dans des calculs ne se rapportant pas aux mêmes choses,
-# mais si les start-end fonctionne ça ne devrait pas avoir d'impact sur les calculs ? qu'en penses-tu ?
-
-
-class f3vl(Variable):
-    cerfa_field = u"3VL"
+class f3tz(Variable):
+    cerfa_field = u"3TZ"
     value_type = int
     unit = 'currency'
     entity = FoyerFiscal
-    label = u"Distributions par des sociétés de capital-risque taxables à 19 %"
-    end = '2013-12-31'
+    label = u"Plus-values de cession de titres d'OPC monétaires en report d'imposition, plus-values réalisées du 1.1 au 31.3.2017, plus-values en report d'imposition"
+    # start_date = date(2016, 1, 1)
     definition_period = YEAR
 
 
-class f3vi(Variable):
-    cerfa_field = {0: u"3VI",
-        1: u"3SI",
-        }
-    value_type = int
-    unit = 'currency'
-    entity = Individu
-    label = u"Gains de levée d'options sur titres et gains d'acquisition d'actions taxables à 30 %"
-    definition_period = YEAR
-
-  # (f3vi, f3si )
-
-class f3vm(Variable):
-    cerfa_field = u"3VM"
+class f3sb(Variable):
+    cerfa_field = u"3SB"
     value_type = int
     unit = 'currency'
     entity = FoyerFiscal
-    label = u"Clôture du PEA avant l'expiration de la 2e année: gains taxables à 22.5 %"
+    label = u"Plus-values en report d'imposition, dont le report a expiré cette année"
+    # start_date = date(2013, 1, 1)
     definition_period = YEAR
 
 
-class f3vt(Variable):
-    cerfa_field = u"3VT"
+class f3wb(Variable):
+    cerfa_field = u"3WB"
     value_type = int
     unit = 'currency'
     entity = FoyerFiscal
-    label = u"Clôture du PEA  entre la 2e et la 5e année: gains taxables à 19 %"
-    # start_date = date(2012, 1, 1)
+    label = u"Plus-values des individus transférant leur domicile fiscal hors de France; plus-values et créances ne bénéficiant pas du sursis de paiement; plus-values nettes imposables au barème"
+    # start_date = date(2013, 1, 1)
     definition_period = YEAR
 
 
-class f3vj(Variable):
-    cerfa_field = {0: u"3VJ",
-        1: u"3VK",
-        }
+class f3wd(Variable):
+    cerfa_field = u"3WD"
     value_type = int
     unit = 'currency'
-    entity = Individu
-    label = u"Gains imposables sur option dans la catégorie des salaires"
+    entity = FoyerFiscal
+    label = u"Plus-values des individus transférant leur domicile fiscal hors de France; plus-values et créances ne bénéficiant pas du sursis de paiement; plus-values de base soumise aux prélèvements sociaux"
+    # start_date = date(2013, 1, 1)
     definition_period = YEAR
 
-  # (f3vj, f3vk )
 
-class f3va(Variable):
+class f3we(Variable):
+    cerfa_field = u"3WE"
+    value_type = int
+    unit = 'currency'
+    entity = FoyerFiscal
+    label = u"Plus-values en report d'imposition : complément de prix perçu cette année"
+    definition_period = YEAR
+
+
+class f3wn(Variable):
+    cerfa_field = u"3WN"
+    value_type = int
+    unit = 'currency'
+    entity = FoyerFiscal
+    label = u"Plus-values en report d'imposition dont le report a expiré cette année, réalisées à compter du 1.1.2013 : plus-values avant abattement"
+    # start_date = date(2016, 1, 1)
+    definition_period = YEAR
+
+
+class f3wp(Variable):
+    cerfa_field = u"3WP"
+    value_type = int
+    unit = 'currency'
+    entity = FoyerFiscal    
+    label = u"Plus-values en report d'imposition dont le report a expiré cette année, réalisées à compter du 1.1.2013 : plus-values après abattement"
+    # start_date = date(2016, 1, 1)
+    definition_period = YEAR
+
+
+class f3wr(Variable):
+    cerfa_field = u"3WR"
+    value_type = int
+    unit = 'currency'
+    entity = FoyerFiscal
+    label = u"Plus-values en report d'imposition dont le report a expiré cette année, réalisées à compter du 1.1.2013 : impôt sur le revenu"
+    # start_date = date(2016, 1, 1)
+    definition_period = YEAR
+
+
+class f3wt(Variable):
+    cerfa_field = u"3WT"
+    value_type = int
+    unit = 'currency'
+    entity = FoyerFiscal
+    label = u"Plus-values en report d'imposition dont le report a expiré cette année, réalisées à compter du 1.1.2013 : contribution exceptionnelle sur les hauts revenus"
+    # start_date = date(2016, 1, 1)
+    definition_period = YEAR
+
+
+# Abattements sur plus-values
+
+class f3va_2014(Variable):
     cerfa_field = u"3VA"
     value_type = int
     unit = 'currency'
     entity = FoyerFiscal
-    label = u""
+    label = u"Abattements nets (abattement pour durée de détention renforcé et abattement fixe spécial) appliqués sur des plus-values réalisées par les dirigeants de PME lors de leur départ à la retraite"
     # start_date = date(2006, 1, 1)
+    end = '2014-12-31'
     definition_period = YEAR
+
+
+class f3va(Variable):
+    cerfa_field = {
+        0: u"3VA",
+        1: u"3VB",
+        2: u"3VO",
+        3: u"3VP",
+        }
+    value_type = int
+    unit = 'currency'
+    entity = Individu
+    label = u"Abattements nets (abattement pour durée de détention renforcé et abattement fixe spécial) appliqués sur des plus-values réalisées par les dirigeants de PME lors de leur départ à la retraite"
+    # start_date = date(2015, 1, 1)
+    definition_period = YEAR
+
 
 class f3vb(Variable):
     cerfa_field = u"3VB"
     value_type = int
     unit = 'currency'
     entity = FoyerFiscal
-    label = u""
+    label = u"Abattements nets (abattement pour durée de détention renforcé et abattement fixe spécial) appliqués sur des moins-values réalisées par les dirigeants de PME lors de leur départ à la retraite"
     # start_date = date(2006, 1, 1)
+    end = '2014-12-31'
     definition_period = YEAR
+
 
 class f3sg(Variable):
     cerfa_field = u"3SG"
@@ -236,6 +322,7 @@ class f3sg(Variable):
     label = u"Abattement net pour durée de détention : appliqué sur des plus-values"
     # start_date = date(2013, 1, 1)
     definition_period = YEAR
+
 
 class f3sh(Variable):
     cerfa_field = u"3SH"
@@ -247,6 +334,7 @@ class f3sh(Variable):
     end = '2014-12-31'
     definition_period = YEAR
 
+
 class f3sl(Variable):
     cerfa_field = u"3SL"
     value_type = int
@@ -255,6 +343,7 @@ class f3sl(Variable):
     label = u"Abattement net pour durée de détention renforcée : appliqué sur des plus-values"
     # start_date = date(2013, 1, 1)
     definition_period = YEAR
+
 
 class f3sm(Variable):
     cerfa_field = u"3SM"
@@ -276,39 +365,23 @@ class abattement_net_retraite_dirigeant_pme(Variable):
     definition_period = YEAR
 
     def formula_2006_01_01(foyer_fiscal, period):
-        f3va = foyer_fiscal('f3va', period)
-        f3vb = foyer_fiscal('f3va', period)
+        f3va = foyer_fiscal('f3va_2014', period)
+        f3vb = foyer_fiscal('f3vb', period)
 
         return f3va - f3vb
 
-class abatnet_duree_detention(Variable):
-    value_type = int
-    unit = 'currency'
-    entity = FoyerFiscal
-    label = u"Abattement net pour durée de détention"
-    definition_period = YEAR
-
-    def formula_2013_01_01(foyer_fiscal, period):
-        f3sg = foyer_fiscal('f3sg', period)
-        f3sh = foyer_fiscal('f3sh', period)
-        f3sl = foyer_fiscal('f3sl', period)
-        f3sm = foyer_fiscal('f3sm', period)
-
-        return max_(0, f3sg - f3sh) + max_(0, f3sl - f3sm)
-
     def formula_2015_01_01(foyer_fiscal, period):
-        f3sg = foyer_fiscal('f3sg', period)
-        f3sl = foyer_fiscal('f3sl', period)
+        f3va_i = foyer_fiscal.members('f3va', period)
+        f3va = foyer_fiscal.sum(f3va_i)
 
-        return f3sg + f3sl
-
+        return f3va
 
 class abattement_net_duree_detention(Variable):
     value_type = int
     unit = 'currency'
     entity = FoyerFiscal
     reference = u"http://bofip.impots.gouv.fr/bofip/9540-PGP"
-    label = u"Abattement net pour durée de détention"
+    label = u"Abattement net pour durée de détention (de droit commun et renforcé)"
     definition_period = YEAR
 
     def formula_2013_01_01(foyer_fiscal, period):
@@ -328,42 +401,122 @@ class abattement_net_duree_detention(Variable):
 
 # Plus values et gains taxables à des taux forfaitaires
 
-class f3vg(Variable):
-    cerfa_field = u"3VG"
+class f3vd(Variable):
+    cerfa_field = {
+        0: u"3VD",
+        1: u"3SD",
+        }
+    value_type = int
+    unit = 'currency'
+    entity = Individu
+    label = u"Gains de levée d'options sur titres et gains d'acquisition d'actions taxables à 18 %"
+    # start_date = date(2008, 1, 1)
+    definition_period = YEAR
+
+
+class f3vi(Variable):
+    cerfa_field = {
+        0: u"3VI",
+        1: u"3SI",
+        }
+    value_type = int
+    unit = 'currency'
+    entity = Individu
+    label = u"Gains de levée d'options sur titres et gains d'acquisition d'actions taxables à 30 %"
+    definition_period = YEAR
+
+
+class f3vf(Variable):
+    cerfa_field = {
+        0: u"3VF",
+        1: u"3SF",
+        }
+    value_type = int
+    unit = 'currency'
+    entity = Individu
+    label = u"Gains de levée d'options sur titres et gains d'acquisition d'actions taxables à 41 %"
+    definition_period = YEAR
+
+
+class f3vm(Variable):
+    cerfa_field = u"3VM"
     value_type = int
     unit = 'currency'
     entity = FoyerFiscal
-    label = u"Plus-value imposable sur gains de cession de valeurs mobilières, de droits sociaux et gains assimilés"
+    label = u"Clôture du PEA avant l'expiration de la 2e année: gains taxables à 22.5 %"
+    # start_date = date(2012, 1, 1)
     definition_period = YEAR
 
 
-class f3vh(Variable):
-    cerfa_field = u"3VH"
+class f3vt(Variable):
+    cerfa_field = u"3VT"
     value_type = int
     unit = 'currency'
     entity = FoyerFiscal
-    label = u"Perte de l'année de perception des revenus"
+    label = u"Clôture du PEA  entre la 2e et la 5e année: gains taxables à 19 %"
+    # start_date = date(2012, 1, 1)
     definition_period = YEAR
 
 
-class f3vu(Variable):
-    value_type = int
-    entity = FoyerFiscal
-    end = '2009-12-31'
-    definition_period = YEAR
-
-  # TODO: vérifier pour 2010 et 2011
-
-class f3vv(Variable):
-    cerfa_field = u"3VV"
+class f3vl(Variable):
+    cerfa_field = u"3VL"
     value_type = int
     unit = 'currency'
     entity = FoyerFiscal
-    label = u"Plus-values réalisées par les non-résidents: montant du prélèvement de 45 % déjà versé"
-    # start_date = date(2013, 1, 1)
+    label = u"Distributions par des sociétés de capital-risque taxables à 19 %"
+    end = '2013-12-31'
     definition_period = YEAR
 
-  # TODO: à revoir :ok pour 2013, pas de 3vv pour 2012, et correspond à autre chose en 2009, vérifier 2010 et 2011
+
+class f3sa(Variable):
+    cerfa_field = u"3SA"
+    value_type = int
+    entity = FoyerFiscal
+    label = u"Plus-values de cessions de titres réalisées par un entrepreneur, taxables à 19%"
+    # start_date = date(2012, 1, 1)
+    end = '2012-12-31'
+    definition_period = YEAR
+
+
+class f3sj(Variable):
+    cerfa_field = u"3SJ"
+    value_type = int
+    unit = 'currency'
+    entity = FoyerFiscal
+    label = u"Gains de cession de bons de souscription de parts de créateur d'entreprise : gains taxables à 19%"
+    definition_period = YEAR
+
+
+class f3sk(Variable):
+    cerfa_field = u"3SK"
+    value_type = int
+    unit = 'currency'
+    entity = FoyerFiscal
+    label = u"Gains de cession de bons de souscription de parts de créateur d'entreprise : gains taxables à 30%"
+    definition_period = YEAR
+
+
+class f3wi(Variable):
+    cerfa_field = u"3WI"
+    value_type = int
+    unit = 'currency'
+    entity = FoyerFiscal
+    label = u"Plus-values en report d'imposition dont le report a expiré cette année : réalisées du 14.11.2012 au 31.12.2012 et taxables à 24%"
+    # start_date = date(2016, 1, 1)
+    definition_period = YEAR
+
+
+class f3wj(Variable):
+    cerfa_field = u"3WJ"
+    value_type = int
+    unit = 'currency'
+    entity = FoyerFiscal
+    label = u"Plus-values en report d'imposition dont le report a expiré cette année : réalisées du 14.11.2012 au 31.12.2012 et taxables à 19%"
+    # start_date = date(2016, 1, 1)
+    definition_period = YEAR
+
+
+# Autres variables
 
 class f3vv_end_2010(Variable):
     cerfa_field = u"3VV"
@@ -376,13 +529,8 @@ class f3vv_end_2010(Variable):
     definition_period = YEAR
 
 
-class f3vz(Variable):
-    cerfa_field = u"3VZ"
+class f3vu(Variable):
     value_type = int
-    unit = 'currency'
     entity = FoyerFiscal
-    label = u"Plus-values imposables sur cessions d’immeubles ou de biens meubles"
-    # start_date = date(2011, 1, 1)
+    end = '2009-12-31'
     definition_period = YEAR
-
-  # TODO: vérifier avant 2012